Move Git Integration Tests and code review fixes
