+++ tests/unit_and_functional/test_git.py
@@ -32,56 +32,20 @@ from bodywork.git import (
     download_project_code_from_repo,
     get_connection_protocol,
     setup_ssh_for_git_host,
-    get_ssh_public_key_from_domain,
+    get_ssh_public_key_from_domain
 )
 
 
 def test_that_git_project_repo_can_be_cloned(
     setup_bodywork_test_project: Iterable[bool],
     project_repo_connection_string: str,
-    cloned_project_repo_location: Path,
+    cloned_project_repo_location: Path
 ):
     try:
         download_project_code_from_repo(project_repo_connection_string)
         assert cloned_project_repo_location.exists()
     except Exception:
         assert False
-
-
-def test_that_git_project_repo_can_be_cloned_from_github_using_ssh(
-    setup_bodywork_test_project: Iterable[bool],
-    github_repo_connection_string: str,
-    cloned_project_repo_location: Path,
-    set_github_ssh_private_key_env_var: None,
-):
-    try:
-        download_project_code_from_repo(github_repo_connection_string)
-        assert cloned_project_repo_location.exists()
-    except Exception:
-        assert False
-    finally:
-        ssh_dir = Path(".") / SSH_DIR_NAME
-        shutil.rmtree(ssh_dir, ignore_errors=True, onerror=on_error)
-
-
-def test_that_git_project_repo_can_be_cloned_from_gitlab_using_ssh(
-    setup_bodywork_test_project: Iterable[bool],
-    gitlab_repo_connection_string: str,
-    cloned_project_repo_location: Path,
-    set_gitlab_ssh_private_key_env_var: None,
-):
-    try:
-        ssh_dir = Path(".") / SSH_DIR_NAME
-        os.makedirs(ssh_dir, exist_ok=False)
-        with open(os.path.join(ssh_dir, "id_rsa"), "w"):
-            pass
-        download_project_code_from_repo(gitlab_repo_connection_string)
-        assert cloned_project_repo_location.exists()
-    except Exception:
-        assert False
-    finally:
-        ssh_dir = Path(".") / SSH_DIR_NAME
-        shutil.rmtree(ssh_dir, ignore_errors=True, onerror=on_error)
 
 
 def test_that_git_project_clone_raises_exceptions():
@@ -126,8 +90,8 @@ def test_setup_ssh_for_git_host_create_known_host_and_env_var():
     if not os.environ.get(SSH_PRIVATE_KEY_ENV_VAR):
         os.environ[SSH_PRIVATE_KEY_ENV_VAR] = "MY_PRIVATE_KEY"
 
+    ssh_dir = Path(".") / SSH_DIR_NAME
     try:
-        ssh_dir = Path(".") / SSH_DIR_NAME
         assert ssh_dir.exists() is False
         setup_ssh_for_git_host("github.com")
